==> 04_quarto.qmd <==
---
title: 'Wastewater-based surveillance report'
subtitle: "Respiratory viruses"
author: "Author: Team 1 (Elisa, Gizem, Patrick, Sophie, Kim)"
format:  
  html: 
    toc: true
    # css: style.css
    # page-layout: full
  docx:
    toc: true
editor: source
crossref:
  fig-title: '**Figure**'
  fig-labels: arabic
  title-delim: "**:**"
execute:
  echo: false
  warning: false
# bibliography: bibli.bib
# csl: style.csl
---

```{r}
#| label: load packages and data
load(".RData")

library(flextable)
library(dplyr)

tbl_sars <- readRDS("tbl_sars.RDS")
tbl_brussels_north <- readRDS("site_specific_table.rds")
tbl_rsv <- readRDS("tbl_rsv.RDS")
tbl_flu <- readRDS("tbl_flu.RDS")
```

## Intro

Latest results dating of `r format(date_reporting, "%G-W%V")` (`r format(date_reporting, "%d-%m-%Y")`):

-   `r main_text %>% filter(date_week == format(date_reporting, "%G-W%V")) %>% pull(SARS_nation_level_en)`

## Preamble

This report represents a national overview of wastewater surveillance in Belgium. This work helps understand the underlying concentration and distribution of respiratory pathogen genomes being shed by the population contained within the catchment areas. This work would not be possible without the collaboration of our partners who we thank for their contributions.

## National level

Wastewater surveillance is performed at X number of sites around Belgium. These are sampled and analysed by participating laboratories. With analysis and interpretation provided by the team at Y. Values are normalised and a rolling mean is presented.

```{r}
#| fig-cap: Viral concentration measured in wastewater from two wastewater treatment plants. Normalised SARS-CoV-2 RNA concentrations (SARS-CoV-2/PMMoV) are shown for each sampling date to illustrate trends in wastewater-based surveillance and differences in viral loading across catchments.

knitr::include_graphics("./plot/graph-viral_ratio-nation.png")
```

## Individual area level

The following table shows normalised results at individual sampling locations.

```{r}
#| tbl-cap: Table of the most recent ten national wastewater measurements, sampling dates, and normalised SARS-CoV-2 viral ratios (SARS-CoV-2/PMMoV).
tbl_sars
```

```{r}
#| tbl-cap: Table of the most recent ten wastewater measurements from North Brussels sampling dates, and normalised SARS-CoV-2 viral ratios (SARS-CoV-2/PMMoV).
tbl_brussels_north
```

## Influenza

Following table shows PMMoV normalised results at national level for influenza.
```{r}
#| tbl-cap: Table of the most recent ten national wastewater measurements, sampling dates, and normalised Influenza viral ratios (Influenza/PMMoV).
tbl_flu
```

## RSV

Following table shows PMMoV normalised results at national level for RS-virus.
```{r}
#| tbl-cap: Table of the most recent ten national wastewater measurements, sampling dates, and normalised Influenza viral ratios (RSV/PMMoV).
tbl_rsv
```
